config update

+++ internal/cfg/segment.go
@@ -3,5 +3,5 @@ package cfg
 type Segment struct {
 	MaxIdxSizeByte uint64
 	MaxMsgSizeByte uint64
-	StartOffset    uint32
+	StartOffset    uint64
 }
